--- lib/assets/chart_builder/main.css
@@ -1,169 +1,0 @@
-.app {
-  font-family: "Inter";
-
-  box-sizing: border-box;
-
-  --gray-50: #f8fafc;
-  --gray-100: #f0f5f9;
-  --gray-200: #e1e8f0;
-  --gray-300: #cad5e0;
-  --gray-400: #91a4b7;
-  --gray-500: #61758a;
-  --gray-600: #445668;
-  --gray-800: #1c2a3a;
-
-  --blue-100: #ecf0ff;
-}
-
-.container {
-  border: solid 1px var(--gray-300);
-  border-bottom: solid 1px var(--gray-200);
-  border-radius: 0.5rem 0.5rem 0 0;
-  background-color: rgba(248, 250, 252, 0.3);
-}
-
-.row {
-  display: flex;
-}
-
-.root {
-  display: flex;
-  background-color: var(--blue-100);
-  padding: 8px 16px;
-  margin-bottom: 12px;
-  border-radius: 0.5rem 0.5rem 0 0;
-  border-bottom: solid 1px var(--gray-200);
-}
-
-.input {
-  appearance: none;
-  padding: 8px 12px;
-  background-color: var(--gray-50);
-  font-size: 0.875rem;
-  border: 1px solid var(--gray-200);
-  border-radius: 0.5rem;
-  color: var(--gray-600);
-}
-
-input[type=number] {
-  appearance: textfield;
-}
-
-.input::placeholder {
-  color: var(--gray-400);
-}
-
-.input:focus {
-  outline: none;
-  border: 1px solid var(--gray-300);
-}
-
-.input:disabled {
-  opacity: 50%;
-}
-
-.input-label {
-  display: block;
-  margin-bottom: 2px;
-  font-size: 0.875rem;
-  color: var(--gray-800);
-  font-weight: 500;
-}
-
-.input--xs {
-  width: 30%;
-}
-
-.input--md {
-  width: 58%;
-}
-
-.root .input-label {
-  color: var(--gray-600);
-  padding-right: 6px;
-  font-size: 0.875rem;
-  text-transform: uppercase;
-  padding-top: 2px;
-}
-
-.field {
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 16px;
-  padding: 0 12px;
-  flex-basis: 26%;
-}
-
-.field:first-child {
-  flex-basis: 39%;
-  margin-left: 4px;
-}
-
-.root-field {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: end;
-}
-
-.root-field:first-child {
-  width: 80%;
-  justify-content: start;
-  flex-shrink: 3;
-}
-
-.info-box {
-  margin-bottom: 24px;
-  padding: 16px;
-  border-radius: 0.5rem;
-  white-space: pre-wrap;
-  font-weight: 500;
-  font-size: 0.875rem;
-  background-color: var(--gray-100);
-  color: var(--gray-500);
-}
-
-.info-box p {
-  margin: 0;
-  padding: 1em 0 0.3em;
-}
-
-.info-box p:first-child {
-  padding-top: 0;
-}
-
-.info-box span {
-  color: var(--gray-600);
-  padding-left: 0.5em;
-}
-
-.hidden {
-  display: none;
-}
-
-select.input {
-  appearance: none;
-  background-image: url("data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMTIiIGhlaWdodD0iOCIgdmlld0JveD0iMCAwIDEyIDgiIGZpbGw9Im5vbmUiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyI+CjxwYXRoIGQ9Ik01Ljk5OTg5IDQuOTc2NzFMMTAuMTI0OSAwLjg1MTcwOEwxMS4zMDMyIDIuMDMwMDRMNS45OTk4OSA3LjMzMzM3TDAuNjk2NTU1IDIuMDMwMDRMMS44NzQ4OSAwLjg1MTcwOEw1Ljk5OTg5IDQuOTc2NzFaIiBmaWxsPSIjNjE3NThBIi8+Cjwvc3ZnPgo=");
-  background-repeat: no-repeat;
-  background-position: center right 10px;
-  background-size: 10px;
-  padding-right: 28px;
-}
-
-.unavailable-option {
-  color: rgba(0, 0, 0, 0.5);
-}
-
-.unavailable {
-  color: rgba(0, 0, 0, 0.5);
-}
-
-select option {
-  color: black;
-}
-
-input::-webkit-outer-spin-button,
-input::-webkit-inner-spin-button {
-  -webkit-appearance: none;
-  margin: 0;
-}
