+++ mix.exs
@@ -1,7 +1,7 @@
 defmodule Kino.MixProject do
   use Mix.Project
 
-  @version "0.5.2"
+  @version "0.6.0-dev"
   @description "Interactive widgets for Livebook"
 
   def project do
@@ -15,8 +15,7 @@ defmodule Kino.MixProject do
       start_permanent: Mix.env() == :prod,
       deps: deps(),
       docs: docs(),
-      package: package(),
-      xref: [exclude: [VegaLite, Ecto.Query, Ecto.Queryable, DBConnection]]
+      package: package()
     ]
   end
 
@@ -33,30 +32,23 @@ defmodule Kino.MixProject do
   defp deps do
     [
       {:table, "~> 0.1.0"},
-      {:vega_lite, "~> 0.1.0", optional: true},
-      {:ecto, "~> 3.0", optional: true},
-      {:postgrex, "~> 0.16", optional: true},
-      {:myxql, "~> 0.6", optional: true},
-      {:db_connection, "~> 2.4.2", optional: true},
-      {:ex_doc, "~> 0.24", only: :dev, runtime: false}
+      {:ex_doc, "~> 0.28", only: :dev, runtime: false}
     ]
   end
 
   defp docs do
     [
       main: "Kino",
-      source_url: "https://github.com/elixir-nx/kino",
+      source_url: "https://github.com/livebook-dev/kino",
       source_ref: "v#{@version}",
       logo: "images/kino_without_text.png",
       groups_for_modules: [
         Widgets: [
           Kino.DataTable,
           Kino.ETS,
-          Kino.Ecto,
           Kino.Frame,
           Kino.Image,
-          Kino.Markdown,
-          Kino.VegaLite
+          Kino.Markdown
         ],
         Inputs: [
           Kino.Input,
@@ -71,6 +63,9 @@ defmodule Kino.MixProject do
         Internal: [
           Kino.Render,
           Kino.Output
+        ],
+        Testing: [
+          KinoTest
         ]
       ]
     ]
@@ -80,7 +75,7 @@ defmodule Kino.MixProject do
     [
       licenses: ["Apache-2.0"],
       links: %{
-        "GitHub" => "https://github.com/elixir-nx/kino"
+        "GitHub" => "https://github.com/livebook-dev/kino"
       }
     ]
   end
